add rota /payments-summary

+++ src/rinhabackend.Application/DTOs/PaymentDto.cs
@@ -3,6 +3,6 @@
 public class PaymentDto()
 {
     public required string correlationId { get; set; }
-    public required decimal amount { get; set; }
+    public required double amount { get; set; }
     public required DateTime date { get; set; }
 };
